+++ HISTORY.rst
@@ -1,3 +1,9 @@
+0.2.0 (2014-03-23)
+++++++++++++++++++
+
+* Added console executable godaddy
+* Added config support via ~/.godaddyrc
+
 0.1.1-0.1.6 (2013-07-12)
 ++++++++++++++++++++++++